Move useSliderVisibility to a separate file
