+++ react/components/SliderTrack.tsx
@@ -1,8 +1,9 @@
-import React, { FC, useEffect, useRef } from 'react'
+import React, { FC } from 'react'
 import { useCssHandles, applyModifiers } from 'vtex.css-handles'
 
 import { useSliderState, useSliderDispatch } from './SliderContext'
 import { useSliderGroupDispatch } from '../SliderLayoutGroup'
+import { useSliderVisibility } from '../hooks/useSliderVisibility'
 
 const CSS_HANDLES = ['sliderTrack', 'slide', 'slideChildrenContainer'] as const
 
@@ -10,25 +11,6 @@ interface Props {
   totalItems: number
   infinite: boolean
   usePagination: boolean
-}
-
-const isSlideVisible = ({
-  index,
-  currentSlide,
-  slidesToShow,
-  totalItems,
-}: {
-  index: number
-  currentSlide: number
-  slidesToShow: number
-  totalItems: number
-}): boolean => {
-  const isClonedSlide = currentSlide < 0 || currentSlide >= totalItems
-
-  return (
-    (index >= currentSlide && index < currentSlide + slidesToShow) ||
-    isClonedSlide
-  )
 }
 
 const resolveAriaAttributes = (
@@ -63,38 +45,6 @@ const getFirstOrLastVisible = (slidesPerPage: number, index: number) => {
   }
 
   return ''
-}
-
-const useSliderVisibility = (
-  currentSlide: number,
-  slidesPerPage: number,
-  totalItems: number
-) => {
-  /** Keeps track of slides that have been visualised before.
-   * We want to keep rendering them because the issue is mostly rendering
-   * slides that might never be viewed; On the other hand, hiding slides
-   * that were visible causes visual glitches */
-  const visitedSlides = useRef<Set<number>>(new Set())
-
-  useEffect(() => {
-    for (let i = 0; i < slidesPerPage; i++) {
-      visitedSlides.current.add(currentSlide + i)
-    }
-  }, [currentSlide, slidesPerPage])
-
-  const isItemVisible = (index: number) =>
-    isSlideVisible({
-      index,
-      currentSlide,
-      slidesToShow: slidesPerPage,
-      totalItems,
-    })
-
-  const shouldRenderItem = (index: number) => {
-    return visitedSlides.current.has(index) || isItemVisible(index)
-  }
-
-  return { shouldRenderItem, isItemVisible }
 }
 
 const SliderTrack: FC<Props> = ({ totalItems, infinite, usePagination }) => {
